code/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
openai>=1.0.0
python-dotenv>=0.19.0

# Document processing
langchain>=0.1.0
docling>=0.1.0  # For PDF processing

# Text processing
tiktoken>=0.5.0

# Optional: GPU acceleration for similarity calculations
cupy-cuda11x>=12.0.0  # For CUDA 11.x (adjust version as needed)
# cupy-cuda12x>=12.0.0  # For CUDA 12.x (adjust version as needed)

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0 